requirements.txt
customtkinter==5.2.2
Pillow==10.3.0
requests==2.31.0
psutil==5.9.8
pyaudio==0.2.14
pystray==0.19.5
pywin32==306
